+++ src/Extractor/SourceTargetMappingExtractor.php
@@ -79,7 +79,7 @@ class SourceTargetMappingExtractor extends MappingExtractor
                     $maxDepth,
                     $this->isIgnoredProperty($mapperMetadata->getSource(), $property),
                     $this->isIgnoredProperty($mapperMetadata->getTarget(), $property),
-                    PropertyReadInfo::VISIBILITY_PUBLIC === $this->readInfoExtractor->getReadInfo($mapperMetadata->getSource(), $property)?->getVisibility() ?? true,
+                    PropertyReadInfo::VISIBILITY_PUBLIC === ($this->readInfoExtractor->getReadInfo($mapperMetadata->getSource(), $property)?->getVisibility() ?? true),
                 );
             }
         }