+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- [GH#11](https://github.com/jolicode/automapper/pull/11) Added phpstan level 5 in CI
 
 ## [8.0.1] - 2023-10-04
 ### Changed