+++ CHANGELOG.md
@@ -9,6 +9,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - [GH#11](https://github.com/jolicode/automapper/pull/11) Added phpstan level 5 in CI
 
+### Fixed
+- [GH#9](https://github.com/jolicode/automapper/pull/9) fix: `mapPrivatePropertiesAndMethod` should not be mandatory
+
 ## [8.0.1] - 2023-10-04
 ### Changed
 - [GH#6](https://github.com/jolicode/automapper/pull/6) Document all AST code by explaining what it generates