+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [8.0.2] - 2023-11-06
 ### Added
 - [GH#11](https://github.com/jolicode/automapper/pull/11) Added phpstan level 5 in CI
 
@@ -176,4 +178,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 * [AutoMapper] [GH#179](https://github.com/janephp/janephp/pull/179) Fixing incompatible changes in Symfony 5.0
 
-[Unreleased]: https://github.com/jolicode/automapper/compare/v7.5.3...HEAD
+[Unreleased]: https://github.com/jolicode/automapper/compare/8.0.2...HEAD
+[8.0.2]: https://github.com/janephp/janephp/compare/8.0.1...8.0.2
+[8.0.1]: https://github.com/janephp/janephp/compare/8.0.0...8.0.1